+++ docs/post_build.html
@@ -28,7 +28,7 @@
   returned values). Having got these, one can then call <code>set_command(name, cmd)</code>
   to alter the command for your rule.</p>
 
-<p>The built-in C++ rules for Please (<code><a href="https://github.com/thought-machine/please/blob/master/src/parse/rules/cc_rules.build_defs">src/parse/rules/cc_rules.build_defs</a></code>) are a reasonably good example of how to use this.</p>
+<p>The built-in C++ rules for Please (<code><a href="https://github.com/thought-machine/please/blob/master/rules/cc_rules.build_defs">rules/cc_rules.build_defs</a></code>) are a reasonably good example of how to use this.</p>
 
 <h2>Post-build function</h2>
