+++ src/parse/BUILD
@@ -12,8 +12,8 @@ go_library(
         "//src/core",
         "//src/fs",
         "//src/parse/asp",
-        "//src/parse/rules",
-        "//src/parse/rules/bazel",
+        "//rules",
+        "//rules/bazel",
         "//src/utils",
         "//src/worker",
         "//third_party/go:logging",
