+++ src/parse/asp/BUILD
@@ -42,7 +42,7 @@ go_test(
     deps = [
         ":asp",
         "//src/core",
-        "//src/parse/rules",
+        "//rules",
         "//third_party/go:testify",
     ],
 )
@@ -54,7 +54,7 @@ go_test(
     deps = [
         ":asp",
         "//src/core",
-        "//src/parse/rules",
+        "//rules",
         "//third_party/go:logging",
         "//third_party/go:testify",
     ],
