+++ src/parse/asp/main/main.go
@@ -20,7 +20,7 @@ import (
 	"github.com/thought-machine/please/src/cli"
 	"github.com/thought-machine/please/src/core"
 	"github.com/thought-machine/please/src/parse/asp"
-	"github.com/thought-machine/please/src/parse/rules"
+	"github.com/thought-machine/please/rules"
 )
 
 var log = logging.MustGetLogger("parser")
@@ -114,7 +114,7 @@ func main() {
 		if strings.HasSuffix(filename, ".gob") {
 			srcFile := strings.TrimSuffix(filename, ".gob")
 			src, _ := rules.Asset(srcFile)
-			p.MustLoadBuiltins("src/parse/rules/"+srcFile, src, rules.MustAsset(filename))
+			p.MustLoadBuiltins("rules/"+srcFile, src, rules.MustAsset(filename))
 		}
 	}
