+++ src/parse/init.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/thought-machine/please/src/core"
 	"github.com/thought-machine/please/src/parse/asp"
-	"github.com/thought-machine/please/src/parse/rules"
-	"github.com/thought-machine/please/src/parse/rules/bazel"
+	"github.com/thought-machine/please/rules"
+	"github.com/thought-machine/please/rules/bazel"
 )
 
 // InitParser initialises the parser engine. This is guaranteed to be called exactly once before any calls to Parse().
@@ -35,7 +35,7 @@ func newAspParser(state *core.BuildState) *asp.Parser {
 		if strings.HasSuffix(filename, ".gob") {
 			srcFile := strings.TrimSuffix(filename, ".gob")
 			src, _ := rules.Asset(srcFile)
-			p.MustLoadBuiltins("src/parse/rules/"+srcFile, src, rules.MustAsset(filename))
+			p.MustLoadBuiltins("rules/"+srcFile, src, rules.MustAsset(filename))
 		}
 	}
