+++ tools/build_langserver/langserver/analyzer_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/thought-machine/please/src/core"
 	"github.com/thought-machine/please/src/parse/asp"
-	"github.com/thought-machine/please/src/parse/rules"
+	"github.com/thought-machine/please/rules"
 	"github.com/thought-machine/please/tools/build_langserver/lsp"
 
 	"github.com/stretchr/testify/assert"
@@ -182,7 +182,7 @@ func TestBuildLabelFromString(t *testing.T) {
 		"        \"//src/fs\",\n" +
 		"        \"//src/parse\",\n" +
 		"        \"//src/parse/asp\",\n" +
-		"        \"//src/parse/rules\",\n" +
+		"        \"//rules\",\n" +
 		"        \"//third_party/go:jsonrpc2\",\n" +
 		"        \"//third_party/go:logging\",\n" +
 		"        \"//tools/build_langserver/lsp\",\n" +
