+++ tools/build_langserver/langserver/test_data/assignment.build
@@ -11,7 +11,7 @@ label_list_multi = [
      "//src/core",
      "//src/parse",
      "//src/parse/asp",
-     "//src/parse/rules",
+     "//rules",
      "//third_party/go:jsonrpc2",
      "//third_party/go:logging",
      "//tools/build_langserver/lsp",
