+++ tools/build_langserver/langserver/test_data/dummy_build_defs/BUILD
@@ -9,7 +9,7 @@ randoms = [
     "//src/fs",
     "//src/parse",
     "//src/parse/asp",
-    "//src/parse/rules",
+    "//rules",
     "//third_party/go:jsonrpc2",
     "//third_party/go:logging",
     "//tools/build_langserver/lsp",
