+++ tools/build_langserver/langserver/test_data/example2.build
@@ -12,7 +12,7 @@ go_library(
         "//src/fs",
         "//src/parse",
         "//src/parse/asp",
-        "//src/parse/rules",
+        "//rules",
         "//third_party/go:jsonrpc2",
         "//third_party/go:logging",
         "//tools/build_langserver/lsp",
@@ -28,7 +28,7 @@ go_test(
     data = ["test_data"],
     deps = [
         ":langserver",
-        "//src/parse/rules",
+        "//rules",
         "//third_party/go:testify",
     ],
 )
